Move sig related functions to its own module
