+++ dynosaur_derive/src/expand.rs
@@ -1,5 +1,6 @@
 use crate::lifetime::{used_lifetimes, AddLifetimeToImplTrait, CollectLifetimes};
 use crate::receiver::has_self_in_sig;
+use crate::sig::{is_async, is_rpit};
 use crate::where_clauses::{has_where_self_sized, where_clause_or_default};
 use proc_macro2::{Span, TokenStream};
 use quote::quote;
@@ -49,56 +50,6 @@ pub(crate) fn expand_fn_sig(item_trait_generics: &Generics, trait_item_fn: &mut 
 
     // Remove default method if any for the erased trait
     trait_item_fn.default = None;
-}
-
-pub(crate) fn is_async(sig: &Signature) -> bool {
-    match sig {
-        Signature {
-            asyncness: Some(_), ..
-        } => true,
-        Signature {
-            asyncness: None,
-            output: ReturnType::Type(_, ret),
-            ..
-        } => {
-            if let Type::ImplTrait(type_impl_trait) = &**ret {
-                type_impl_trait
-                    .bounds
-                    .iter()
-                    .find(|bound| match bound {
-                        TypeParamBound::Trait(trait_bound) => {
-                            let segments = &trait_bound.path.segments;
-
-                            segments.len() == 3
-                                && (segments[0].ident == "core" || segments[0].ident == "std")
-                                && segments[1].ident == "future"
-                                && segments[2].ident == "Future"
-                        }
-                        _ => false,
-                    })
-                    .is_some()
-            } else {
-                false
-            }
-        }
-        _ => false,
-    }
-}
-
-pub(crate) fn is_rpit(sig: &Signature) -> bool {
-    match sig {
-        Signature {
-            asyncness: Some(_), ..
-        } => true,
-        Signature {
-            asyncness: None,
-            output: ReturnType::Type(_, ret),
-            ..
-        } => {
-            matches!(**ret, Type::ImplTrait(_))
-        }
-        _ => false,
-    }
 }
 
 fn expand_fn_input(item_trait_generics: &Generics, sig: &mut Signature) {
